Refactor system services into focused files
