+++ Sources/Capsomnia/ClamshellStateReader.swift
@@ -1,0 +1,25 @@
+import Foundation
+import IOKit
+
+enum ClamshellStateReader {
+    static func isClosed() -> Bool? {
+        let service = IOServiceGetMatchingService(kIOMainPortDefault, IOServiceMatching("IOPMrootDomain"))
+        guard service != 0 else { return nil }
+        defer { IOObjectRelease(service) }
+
+        guard let value = IORegistryEntryCreateCFProperty(
+            service,
+            "AppleClamshellState" as CFString,
+            kCFAllocatorDefault,
+            0
+        )?.takeRetainedValue() else {
+            return nil
+        }
+
+        if let boolValue = value as? Bool {
+            return boolValue
+        }
+
+        return (value as? NSNumber)?.boolValue
+    }
+}
