+++ Sources/Capsomnia/DisplaySleepPolicy.swift
@@ -1,0 +1,5 @@
+enum DisplaySleepPolicy {
+    static func shouldRequestDisplaySleep(externalDisplayConnected: Bool?) -> Bool {
+        externalDisplayConnected == false
+    }
+}
