+++ Sources/Capsomnia/SleepStateReader.swift
@@ -1,0 +1,25 @@
+enum SleepStateReader {
+    static func isDisabled() -> Bool? {
+        let result = CommandRunner.run("/usr/bin/pmset", ["-g"])
+        guard result.status == 0 else { return nil }
+        return parse(result.stdout)
+    }
+
+    static func parse(_ output: String) -> Bool? {
+        for line in output.split(whereSeparator: { $0.isNewline }) {
+            let fields = line.split(whereSeparator: { $0.isWhitespace })
+            guard fields.count >= 2,
+                  fields[0].lowercased() == "sleepdisabled" else {
+                continue
+            }
+
+            switch fields[1] {
+            case "1": return true
+            case "0": return false
+            default: return nil
+            }
+        }
+
+        return nil
+    }
+}
