--- Sources/Capsomnia/SystemServices.swift
@@ -1,210 +1,0 @@
-import CoreGraphics
-import Foundation
-import IOKit
-
-typealias CommandResult = (status: Int32, stdout: String, stderr: String)
-
-struct LaunchAgentError: LocalizedError {
-    let message: String
-
-    var errorDescription: String? {
-        message
-    }
-}
-
-private final class CommandOutputBuffer: @unchecked Sendable {
-    private let lock = NSLock()
-    private var data = Data()
-
-    func append(_ newData: Data) {
-        lock.lock()
-        defer { lock.unlock() }
-        data.append(newData)
-    }
-
-    func string() -> String {
-        lock.lock()
-        defer { lock.unlock() }
-        return String(data: data, encoding: .utf8)?
-            .trimmingCharacters(in: .whitespacesAndNewlines) ?? ""
-    }
-}
-
-enum CommandRunner {
-    static func run(_ executablePath: String, _ arguments: [String]) -> CommandResult {
-        let process = Process()
-        let stdoutPipe = Pipe()
-        let stderrPipe = Pipe()
-        let stdoutBuffer = CommandOutputBuffer()
-        let stderrBuffer = CommandOutputBuffer()
-        let outputGroup = DispatchGroup()
-        let terminationSemaphore = DispatchSemaphore(value: 0)
-
-        process.executableURL = URL(fileURLWithPath: executablePath)
-        process.arguments = arguments
-        process.standardOutput = stdoutPipe
-        process.standardError = stderrPipe
-        process.terminationHandler = { _ in
-            terminationSemaphore.signal()
-        }
-
-        drain(stdoutPipe.fileHandleForReading, into: stdoutBuffer, group: outputGroup)
-        drain(stderrPipe.fileHandleForReading, into: stderrBuffer, group: outputGroup)
-
-        do {
-            try process.run()
-        } catch {
-            stdoutPipe.fileHandleForWriting.closeFile()
-            stderrPipe.fileHandleForWriting.closeFile()
-            outputGroup.wait()
-            return (-1, "", "\(error)")
-        }
-
-        terminationSemaphore.wait()
-        outputGroup.wait()
-
-        return (
-            process.terminationStatus,
-            stdoutBuffer.string(),
-            stderrBuffer.string()
-        )
-    }
-
-    private static func drain(
-        _ handle: FileHandle,
-        into buffer: CommandOutputBuffer,
-        group: DispatchGroup
-    ) {
-        group.enter()
-        DispatchQueue.global(qos: .utility).async {
-            defer { group.leave() }
-            buffer.append(handle.readDataToEndOfFile())
-        }
-    }
-}
-
-enum SystemSleepRequester {
-    typealias Runner = (String, [String]) -> CommandResult
-
-    static func request() -> CommandResult {
-        request(runner: CommandRunner.run)
-    }
-
-    static func request(runner: Runner) -> CommandResult {
-        runner("/usr/bin/pmset", ["sleepnow"])
-    }
-}
-
-enum LaunchAgentManager {
-    static func setEnabled(_ enabled: Bool) throws {
-        let arguments = [
-            enabled ? "enable" : "disable",
-            "gui/\(getuid())/\(appLabel)"
-        ]
-        let result = CommandRunner.run("/bin/launchctl", arguments)
-        guard result.status == 0 else {
-            throw LaunchAgentError(
-                message: "launchctl \(arguments.joined(separator: " ")) failed: \(result.stderr.isEmpty ? result.stdout : result.stderr)"
-            )
-        }
-    }
-}
-
-enum SleepStateReader {
-    static func isDisabled() -> Bool? {
-        let result = CommandRunner.run("/usr/bin/pmset", ["-g"])
-        guard result.status == 0 else { return nil }
-        return parse(result.stdout)
-    }
-
-    static func parse(_ output: String) -> Bool? {
-        for line in output.split(whereSeparator: { $0.isNewline }) {
-            let fields = line.split(whereSeparator: { $0.isWhitespace })
-            guard fields.count >= 2,
-                  fields[0].lowercased() == "sleepdisabled" else {
-                continue
-            }
-
-            switch fields[1] {
-            case "1": return true
-            case "0": return false
-            default: return nil
-            }
-        }
-
-        return nil
-    }
-}
-
-enum ClamshellStateReader {
-    static func isClosed() -> Bool? {
-        let service = IOServiceGetMatchingService(kIOMainPortDefault, IOServiceMatching("IOPMrootDomain"))
-        guard service != 0 else { return nil }
-        defer { IOObjectRelease(service) }
-
-        guard let value = IORegistryEntryCreateCFProperty(
-            service,
-            "AppleClamshellState" as CFString,
-            kCFAllocatorDefault,
-            0
-        )?.takeRetainedValue() else {
-            return nil
-        }
-
-        if let boolValue = value as? Bool {
-            return boolValue
-        }
-
-        return (value as? NSNumber)?.boolValue
-    }
-}
-
-enum ExternalDisplayReader {
-    static func isConnected() -> Bool? {
-        var displayCount: UInt32 = 0
-        guard CGGetOnlineDisplayList(0, nil, &displayCount) == .success else {
-            return nil
-        }
-
-        guard displayCount > 0 else { return false }
-
-        var displays = [CGDirectDisplayID](repeating: 0, count: Int(displayCount))
-        guard CGGetOnlineDisplayList(displayCount, &displays, &displayCount) == .success else {
-            return nil
-        }
-
-        return displays.prefix(Int(displayCount)).contains { CGDisplayIsBuiltin($0) == 0 }
-    }
-}
-
-enum DisplaySleepPolicy {
-    static func shouldRequestDisplaySleep(externalDisplayConnected: Bool?) -> Bool {
-        externalDisplayConnected == false
-    }
-}
-
-enum ExternalCapsLockOffPolicy {
-    /// Decide whether a Caps Lock turn-off should be treated as external and
-    /// re-asserted instead of followed. While the lid is closed the built-in
-    /// keyboard cannot be pressed, so a turn-off observed in that window
-    /// comes from an external source unless the app itself initiated it.
-    ///
-    /// - An elapsed auto-off timer is an explicit turn-off that must proceed
-    ///   to system sleep, so an in-progress auto-off always wins over the
-    ///   guard.
-    /// - `clamshellClosed == nil` (state unavailable) fails open: the
-    ///   turn-off is honored as usual.
-    static func shouldReassert(
-        preferenceEnabled: Bool,
-        sleepPreventionActive: Bool,
-        recentUserAction: Bool,
-        autoOffInProgress: Bool,
-        clamshellClosed: Bool?
-    ) -> Bool {
-        preferenceEnabled
-            && sleepPreventionActive
-            && !recentUserAction
-            && !autoOffInProgress
-            && clamshellClosed == true
-    }
-}
